Fixed Average, NumberDigits and Person classes. Renamed ForCycle class. Add MainPerson class.
