+++ src/lesson_4/Person.java
@@ -72,14 +72,4 @@ public class Person {
                 ", familyName='" + familyName + '\'' +
                 '}';
     }
-
-    public static void main(String[] args) {
-
-        Person person = new Person("Ivan", "Ivanovich", "Ivanov");
-        System.out.println(person);
-
-        person.setAge(20);
-        System.out.println("Год рождения человека: " + person.getPersonBirthYear());
-
-    }
 }
